% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kb-cmd.r
\name{kb_cmd}
\alias{kb_cmd}
\alias{kb_following}
\alias{kb_followers}
\alias{kb_ping}
\title{Execute a keybase command-line command}
\usage{
kb_cmd(cmd, stdout = TRUE, stderr = FALSE)

kb_following()

kb_followers()

kb_ping()
}
\arguments{
\item{cmd}{command to execute}

\item{stdout, stderr}{logical}
}
\description{
Execute a keybase command-line command
}
\references{
\url{https://keybase.io/docs/command_line}
}
