% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.r
\name{kb_discover}
\alias{kb_discover}
\title{Check Keybase membership and retrieve basic user info}
\usage{
kb_discover(twitter, github, hackernews, web, coinbase, key_fingerprint,
  kb_usernames_only = FALSE)
}
\arguments{
\item{twitter, github, hackernews, web, coinbase, key_fingerprint}{vectors of ids/fingerprints
from known public services/data stores. At least one service must have
one character vector populated. There can be multiple items per character
vector.}

\item{kb_usernames_only}{if \code{FALSE}, return a substantial amount of Keybase info/proofs
for each user. If \code{TRUE}, return just the Keybase usernames. Default: \code{FALSE}}
}
\description{
Check Keybase membership and retrieve basic user info
}
\examples{
kb_discover(twitter=c("hrbrmstr", "_inundata", "briandconnelly"), github="bearloga")
kb_discover(twitter=c("sckottie", "_inundata", "briandconnelly"), kb_usernames_only=TRUE)
}
