% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kb-cmd.r
\name{kb_encrypt_file}
\alias{kb_encrypt_file}
\alias{kb_decrypt_file}
\title{Encrypt/decrypt a file}
\usage{
kb_encrypt_file(users, infile, outfile = NULL, binary = FALSE)

kb_decrypt_file(infile, outfile = NULL)
}
\arguments{
\item{users}{keybase user ids (character vector)}

\item{infile}{path to file to encrypt}

\item{outfile}{if \code{NULL} then the encrypted/decrypted file contents are returned as
a vector (type depends on value of \code{binary}). If a string, it should be a
valid path to where the encrypted/decrypted file will be written.}

\item{binary}{if \code{TRUE} iutput in binary (rather than ASCII/armored which is the default)}
}
\description{
Encrypt/decrypt a file
}
