% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kb-cmd.r
\name{kb_encrypt_msg}
\alias{kb_encrypt_msg}
\alias{kb_decrypt_msg}
\title{Encrypt/decrypt a message}
\usage{
kb_encrypt_msg(users, msg)

kb_decrypt_msg(msg)
}
\arguments{
\item{users}{keybase user ids (character vector)}

\item{msg}{message text}
}
\description{
Encrypt/decrypt a message
}
