% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{kb_file_exists}
\alias{kb_file_exists}
\title{See if a file exists}
\usage{
kb_file_exists(user, path)
}
\arguments{
\item{user}{Keybase user ID or a \code{keybase.pub} URL}

\item{path}{path of resource (file) to read (relative from the top level of the public folder).
If \code{user} is a URL this parameter is ignored.}
}
\value{
\code{TRUE} if the specified file exists
}
\description{
See if a file exists
}
\examples{
kb_file_exists("hrbrmstr", "index.md")
kb_file_exists("https://hrbrmstr.keybase.pub/data/Rlogo.pngd")
kb_file_exists("hrbrmstr", "data/Rlogo.png")
}
