% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgp-keys.r
\name{kb_get_public_keys}
\alias{kb_get_public_keys}
\title{Retrive the public key(s) for a given user}
\usage{
kb_get_public_keys(user, import = FALSE)
}
\arguments{
\item{user}{user}

\item{import}{if \code{TRUE}, import into local GPG keyring}
}
\value{
single character vector with the key(s) (invisibly)
}
\description{
Retrive the public key(s) for a given user
}
