% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-files.r
\name{kb_list_files}
\alias{kb_list_files}
\title{List files/directories in a user's public folder}
\usage{
kb_list_files(user, path, pattern = NULL)
}
\arguments{
\item{user}{Keybase user ID or a \code{keybase.pub} URL}

\item{path}{(optional) path (relative from the top level of the public folder).
If \code{user} is a URL this parameter is ignored.}

\item{pattern}{if not \code{NULL}, then the file list will be filtered using the
\code{stringi}-acceptable regular expression passed in}
}
\value{
\code{tibble} with the directory listing results
}
\description{
List files/directories in a user's public folder
}
\examples{
kb_list_files("marcopolo")
kb_list_files("marcopolo", pattern="\\\\.txt")
}
