% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{kb_raw_url}
\alias{kb_raw_url}
\title{Turn a user + path into a Keybase "raw" URL}
\usage{
kb_raw_url(user, path)
}
\arguments{
\item{user}{Keybase user ID}

\item{path}{path of resource (file/dir) relative from the top level of the public folder}
}
\description{
Turn a user + path into a Keybase "raw" URL
}
\examples{
kb_raw_url("hrbrmstr", "data/mtcars.json")
}
