% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{kb_read_file}
\alias{kb_read_file}
\title{Read a resource (file) from a Keybase public folder}
\usage{
kb_read_file(user, path)
}
\arguments{
\item{user}{Keybase user ID or a \code{keybase.pub} URL}

\item{path}{path of resource (file) to read (relative from the top level of the public folder).
If \code{user} is a URL this parameter is ignored.}
}
\value{
the imported object (see \link[rio:import]{rio::import})
}
\description{
Read a resource (file) from a Keybase public folder
}
\examples{
kb_list_files("hrbrmstr", "data")
kb_read_file("hrbrmstr", "data/mtcars.csv")
}
