% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybase-package.R
\docType{package}
\name{keybase}
\alias{keybase}
\alias{keybase-package}
\title{Tools to Work with the 'Keybase' 'API'}
\description{
'Keybase' <keybase.io> is a directory of people and public keys and provides
methods for obtaining public keys, validating users and exchanging files and/or messages
in a secure fashion. Tools are provided to search for and retrieve information about
'Keybase' users, retrieve and import user public keys and list and/or download files.\cr
\cr
There's also a thin but useful R wrapper around many of they \code{keybase} command-line
utility functions.
}
\author{
Ildiko Czeller (czeildi@gmail.com)

Karthik Ram (karthik.ram@gmail.com)

Bob Rudis (bob@rud.is)

Kara Woo
}
