% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{kb_read_rds}
\alias{kb_read_rds}
\alias{kb_write_rds}
\title{Share RDS files with other Keybase users}
\usage{
kb_read_rds(kb_users, path)

kb_write_rds(x, kb_users, path, compress = c("none", "gz", "bz2", "xz"), ...)
}
\arguments{
\item{kb_users}{character vector of (other) Keybase user(s)}

\item{path}{Relative path to read from/write to.}

\item{x}{R object to write to serialise.}

\item{compress}{Compression method to use: "none", "gz" ,"bz", or "xz".}

\item{...}{Additional arguments to connection function. For example, control
the space-time trade-off of different compression methods with
\code{compression}. See \code{\link[=connections]{connections()}} for more details.}
}
\value{
\code{write_rds()} returns \code{x}, invisibly.
}
\description{
Specify the user(s) you want to share an R object with in a character vector
using the Keybase user conventions outline \href{https://keybase.io/docs/kbfs}{in their kbfs introduction}.
You do not need to specify yourself (the function takes care of adding you).
}
\details{
Order of usernames does not matter as Keybase is smart enough to create and
resolve dynamic symbolic links of all user combos.
}
\examples{
\dontrun{
kb_write_rds(mtcars, "marked", "reprex.rds")
kb_read_rds("marked", "reprex.rds")
}
}
